split add students array pages
